refactor: move matching logic into parent policy folder
