+++ staging/src/k8s.io/apiserver/pkg/admission/plugin/policy/validating/controller.go
@@ -38,8 +38,8 @@ import (
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/apiserver/pkg/admission"
 	"k8s.io/apiserver/pkg/admission/plugin/policy/internal/generic"
+	"k8s.io/apiserver/pkg/admission/plugin/policy/matching"
 	celmetrics "k8s.io/apiserver/pkg/admission/plugin/policy/validating/metrics"
-	"k8s.io/apiserver/pkg/admission/plugin/policy/validatingadmissionpolicy/matching"
 	celconfig "k8s.io/apiserver/pkg/apis/cel"
 	"k8s.io/apiserver/pkg/authorization/authorizer"
 	"k8s.io/apiserver/pkg/warning"
